Добавляет модуль modal
