+++ js/main.js
@@ -1,4 +1,3 @@
 import './thumbnails.js';
-import {renderFullSizePicture} from './full-size-picture.js';
-renderFullSizePicture();
+
